Extract `aether_url` crate and canonicalise all paths
