+++ crates/ark_test/Cargo.toml
@@ -14,6 +14,7 @@ rust-version.workspace = true
 workspace = true
 
 [dependencies]
+aether_url.workspace = true
 amalthea.workspace = true
 anyhow.workspace = true
 ark.workspace = true
