+++ crates/oak_db/src/tests/file.rs
@@ -53,7 +53,8 @@ fn test_semantic_index_matches_oak_semantic() {
     let via_salsa = file.semantic_index(&db);
 
     let parse = aether_parser::parse(source, aether_parser::RParserOptions::default());
-    let direct = oak_semantic::build_index(&parse.tree(), &url, oak_semantic::NoopImportsResolver);
+    let direct =
+        oak_semantic::build_index(&parse.tree(), url.as_url(), oak_semantic::NoopImportsResolver);
 
     assert_eq!(via_salsa, &direct);
 }
