+++ crates/oak_db/src/tests/resolver.rs
@@ -37,7 +37,7 @@ fn test_cross_file_source_injection() {
 
     match import_def.unwrap().1.kind() {
         DefinitionKind::Import { file, name, .. } => {
-            assert_eq!(file, b.file(&db).url(&db));
+            assert_eq!(file, b.file(&db).url(&db).as_url());
             assert_eq!(name, "x");
         },
         _ => unreachable!(),
